Bang application forms in component spec and combine system spec
