+++ spec/components/candidate_interface/links_to_previous_applications_component_spec.rb
@@ -2,15 +2,9 @@ require 'rails_helper'
 
 RSpec.describe CandidateInterface::LinksToPreviousApplicationsComponent do
   let(:candidate) { build(:candidate) }
-  let(:application_form) { create(:application_form, candidate: candidate, phase: :apply_2) }
-  let(:previous_application_form) { create(:completed_application_form, candidate: candidate, phase: :apply_2, submitted_at: 3.days.ago) }
-  let(:first_application_form) { create(:completed_application_form, candidate: candidate, phase: :apply_1, submitted_at: 5.days.ago) }
-
-  before do
-    application_form
-    previous_application_form
-    first_application_form
-  end
+  let!(:application_form) { create(:application_form, candidate: candidate, phase: :apply_2) }
+  let!(:previous_application_form) { create(:completed_application_form, candidate: candidate, phase: :apply_2, submitted_at: 3.days.ago) }
+  let!(:first_application_form) { create(:completed_application_form, candidate: candidate, phase: :apply_1, submitted_at: 5.days.ago) }
 
   it 'renders component with links to the candidates previous applications in chronological order' do
     result = render_inline(described_class.new(application_form: application_form))
