+++ spec/system/candidate_interface/candidate_with_unsuccessful_application_applies_again_spec.rb
@@ -18,8 +18,14 @@ RSpec.feature 'Candidate with unsuccessful application' do
     then_i_am_informed_i_can_only_select_one_course
     and_i_can_indeed_only_select_one_course
 
-    when_i_complete_my_application
-    then_my_application_is_submitted
+    then_application_is_submitted
+    and_application_choices_are_complete
+
+    when_i_click_on_the_link_to_my_previous_application
+    then_i_see_the_review_previous_application_page
+
+    when_i_click_back
+    then_i_see_my_current_application_page
   end
 
   def given_the_pilot_is_open
@@ -98,5 +104,21 @@ RSpec.feature 'Candidate with unsuccessful application' do
   def then_my_application_is_submitted
     expect(page).to have_content 'Application successfully submitted'
     expect(ApplicationForm.last.application_choices.first.reload.status).to eq 'application_complete'
+  end
+
+  def when_i_click_on_the_link_to_my_previous_application
+    click_link 'First application'
+  end
+
+  def then_i_see_the_review_previous_application_page
+    expect(page).to have_current_path(candidate_interface_review_previous_application_path(@candidate.id, @application_form.id))
+  end
+
+  def when_i_click_back
+    click_link 'Back'
+  end
+
+  def then_i_see_my_current_application_page
+    expect(page).to have_current_path(candidate_interface_application_form_path)
   end
 end
